pyproject.toml
[project]
name = "oop-chatagent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [{ name = "Muhammad Raffey", email = "muhammadraffey26@gmail.com" }]
requires-python = ">=3.12"
dependencies = [
    "chainlit>=2.4.0",
    "openai-agents>=0.0.3",
    "python-dotenv>=1.0.1",
]


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/oop_chatagent"]
